nit changes and keybord helper class renamed.
