+++ app/src/main/java/org/oppia/app/utility/KeyboardHelper.kt
@@ -4,7 +4,8 @@ import android.content.Context
 import android.view.View
 import android.view.inputmethod.InputMethodManager
 
-class KeyBoardHelpher {
+/** KeyboardHelper help to handle the visiblity of softinputkeybord. */
+class KeyboardHelper {
   companion object {
     fun hideSoftKeyboard(view: View, context: Context) {
       val imm = context.getSystemService(Context.INPUT_METHOD_SERVICE) as InputMethodManager
